% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{from_bbxml_package}
\alias{from_bbxml_package}
\title{Read a BbXML package}
\usage{
from_bbxml_package(path, convert_html_to_markdown = TRUE)
}
\arguments{
\item{path}{The path to the BbXML package}

\item{convert_html_to_markdown}{Whether to convert HTML to Markdown.}
}
\value{
An R2BB object
}
\description{
This function reads a BbXML package zip file and extracts all tests and
pools from it. The results are returned as a list with two attributes:
\code{pools} and \code{tests}, either of which could be an empty list.
}
\details{
By default, rich text fields in HTML are converted to Markdown using
pandoc. pandoc is sufficiently magical that this can work quite well, but
results are not guaranteed to be identical to the original.
}
