% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{from_bbxml_pool}
\alias{from_bbxml_pool}
\title{Convert from Blackboard XML to an R2BB pool}
\usage{
from_bbxml_pool(x, convert_html_to_markdown = TRUE)
}
\arguments{
\item{x}{An xml2 node or nodeset}

\item{convert_html_to_markdown}{Whether to convert HTML to Markdown.}
}
\value{
An R2BB pool
}
\description{
This function converts an object from Blackboard XML format to an R2BB pool.
}
\details{
Generally you will not call this directly; instead, call \code{\link[=from_bbxml_package]{from_bbxml_package()}}
on an exported zip package file.

By default, rich text fields in HTML are converted to Markdown using pandoc. This
can work quite well, but is not perfect; results are not guaranteed to be
identical to the original.
}
\seealso{
\code{\link[=from_bbxml_package]{from_bbxml_package()}}, \code{\link[=from_bbxml_question]{from_bbxml_question()}}, \code{\link[=from_bbxml_test]{from_bbxml_test()}}
}
