% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{from_bbxml_test}
\alias{from_bbxml_test}
\title{Convert from Blackboard XML to an R2BB test}
\usage{
from_bbxml_test(x, convert_html_to_markdown = TRUE, pools_by_name)
}
\arguments{
\item{x}{An xml2 node or nodeset}

\item{convert_html_to_markdown}{Whether to convert HTML to Markdown.}

\item{pools_by_name}{A named list of pools, by resource name; used to fill in
the pool details for random blocks.}
}
\value{
An R2BB test
}
\description{
This function converts an object from Blackboard XML format to an R2BB test.
}
\details{
Generally you will not call this directly; instead, call \code{\link[=from_bbxml_package]{from_bbxml_package()}}
on an exported zip package file.

By default, rich text fields in HTML are converted to Markdown using pandoc. This
can work quite well, but is not perfect; results are not guaranteed to be
identical to the original.
}
\seealso{
\code{\link[=from_bbxml_package]{from_bbxml_package()}}, \code{\link[=from_bbxml_question]{from_bbxml_question()}}, \code{\link[=from_bbxml_pool]{from_bbxml_pool()}}
}
