% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{html_to_md_pandoc}
\alias{html_to_md_pandoc}
\title{Convert an HTML string to markdown using pandoc}
\usage{
html_to_md_pandoc(x)
}
\arguments{
\item{x}{A character vector of HTML text
elements).}
}
\value{
A character vector of markdown text.
}
\description{
This function converts an HTML string to markdown using pandoc. It is a
wrapper around the \code{pandoc_convert} function from the \code{rmarkdown} package.
}
