% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{md_to_html_pandoc}
\alias{md_to_html_pandoc}
\alias{md_to_html_pandoc_convert_deferred}
\title{Convert a markdown string to HTML using pandoc}
\usage{
md_to_html_pandoc(x, deferred = FALSE, ...)

md_to_html_pandoc_convert_deferred(x, ...)
}
\arguments{
\item{x}{A character vector of markdown text (potentially containing multiple
elements), or a string containing text marked for conversion.}

\item{deferred}{Whether to defer conversion of markdown to HTML using pandoc.
If TRUE, the text will be marked for conversion and included in a longer
string or set of strings. Later, these can be converted in one go.}

\item{...}{Additional arguments to pass to \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}.}
}
\value{
A character vector of HTML text.
}
\description{
This function converts a markdown string to HTML using pandoc. It is a
wrapper around \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}.
}
\details{
If \code{deferred} is TRUE, the text will be marked for conversion and included in
a longer string or set of strings. Later, these can be converted in one go
using \code{\link[=md_to_html_pandoc_convert_deferred]{md_to_html_pandoc_convert_deferred()}}.
}
