% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize an r2bb object}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{The object to normalize}

\item{...}{Additional arguments passed to methods}
}
\value{
A normalized r2bb object
}
\description{
This function normalizes an r2bb object. This takes a list, potentially
read from a YAML file, and converts it to the canonical format used by the
package. This also performs validation of the object. This just dispatches
to the appropriate normalize_*() function based on the \verb{$type} field of the
object.
}
\seealso{
\code{\link[=normalize_pool]{normalize_pool()}}, \code{\link[=normalize_test]{normalize_test()}}, \code{\link[=normalize_question]{normalize_question()}}
}
