% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{normalize_pool}
\alias{normalize_pool}
\title{Normalize a question pool}
\usage{
normalize_pool(pool)
}
\arguments{
\item{pool}{The question pool to normalize.}
}
\value{
The normalized question pool
}
\description{
This function normalizes a question pool, which means that it will convert
the pool to the canonical format used by the package. This also performs
validation of the pool.
}
\seealso{
\code{\link[=normalize]{normalize()}}
}
