% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-normalize.R
\name{normalize_question}
\alias{normalize_question}
\title{Normalize a question}
\usage{
normalize_question(question)
}
\arguments{
\item{question}{The question to normalize}
}
\value{
The normalized question
}
\description{
This function normalizes a question to the canonical format used by the
package, which also performs validation of the question.
}
\seealso{
\code{\link[=normalize]{normalize()}}
}
