% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{normalize_test}
\alias{normalize_test}
\title{Normalize a test}
\usage{
normalize_test(test)
}
\arguments{
\item{test}{The test to normalize.}
}
\value{
The normalized test
}
\description{
This function normalizes a test, which means that it will convert
the test to the canonical format used by the package. This also performs
validation of the test.
}
\seealso{
\code{\link[=normalize]{normalize()}}
}
