% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{print.r2bb_pool}
\alias{print.r2bb_pool}
\title{Print a question pool}
\usage{
\method{print}{r2bb_pool}(x, ...)
}
\arguments{
\item{x}{The question pool to print}

\item{...}{Additional arguments passed to methods}
}
\value{
The question pool
}
\description{
This function prints a question pool to the console; it will be converted
to Markdown using \code{\link[=to_markdown.r2bb_pool]{to_markdown.r2bb_pool()}}.
}
\seealso{
\code{\link[=print.r2bb_question]{print.r2bb_question()}}
}
