% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-print.R
\name{print.r2bb_question}
\alias{print.r2bb_question}
\title{Print a question in a human-readable format}
\usage{
\method{print}{r2bb_question}(x, format = c("markdown", "yaml"), ...)
}
\arguments{
\item{x}{The question to print}

\item{format}{The format to print the question in; markdown or yaml}

\item{...}{Additional arguments to pass to the print method}
}
\description{
This function is used to print a question in a human-readable format.
}
