% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{read_pool}
\alias{read_pool}
\title{Read a question pool from a file}
\usage{
read_pool(
  path,
  allow_multiple = FALSE,
  normalize = TRUE,
  normalise = normalize,
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{allow_multiple}{Whether to allow multiple pools in the file.
These must be separate by the YAML document boundary, \verb{---}. If this
is true, the return value is a list of pools; otherwise, the return
value is a single pool.}

\item{normalize}{Whether to normalize the pool. This will convert the
pool to the canonical format used by the package, which also performs
validation of the pool.}

\item{normalise}{Synonym for \code{normalize}.}

\item{envir}{The environment to use for RYaml files.}

\item{...}{Additional arguments to pass to \code{\link[=render_ryaml]{render_ryaml()}}, if it's needed.}

\item{file}{The file to read the question pool from. Either a YAML file, or an
RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used to render it).}
}
\value{
a question pool or a list of question pools
}
\description{
This function reads a question pool from a file. The file can be a YAML
file, or an RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used to render it).
}
