% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-read-write.R
\name{read_question}
\alias{read_question}
\title{Read a question from a file}
\usage{
read_question(
  path,
  allow_multiple = FALSE,
  normalize = TRUE,
  normalise = normalize,
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{path}{The file to read the question from. Either a YAML file, or an
RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used to render it).}

\item{allow_multiple}{Whether to allow multiple questions in the file.
These must be separate by the YAML document boundary, \verb{---}. If this
is true, the return value is a list of questions; otherwise, the return
value is a single question.}

\item{normalize}{Whether to normalize the question. This will convert the
question to the canonical format used by the package, which also performs
validation of the question.}

\item{normalise}{Synonym for \code{normalize}.}

\item{envir}{The environment to use for RYaml files.}

\item{...}{Additional arguments to pass to \code{\link[=render_ryaml]{render_ryaml()}}, if it's needed.}
}
\value{
A question or a list of questions
}
\description{
This function reads a question or a list of questions from a file. The file
can be a YAML file, or an RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used
to render it).
}
