% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryaml.R
\name{read_ryaml}
\alias{read_ryaml}
\title{Read a Ryaml or YAML file}
\usage{
read_ryaml(path, allow_multiple = FALSE, ...)
}
\arguments{
\item{path}{The path to the Ryaml file}

\item{allow_multiple}{Whether to allow multiple questions in the file}

\item{...}{Additional arguments to pass to \code{\link{render_ryaml}}}
}
\value{
If \code{allow_multiple} is \code{FALSE}, the return value is a single R
object. If \code{allow_multiple} is \code{TRUE}, the return value is a list of R
objects.
}
\description{
This function reads an Ryaml or YAML file into an R object. This is almost the
same as calling \code{\link{render_ryaml}} and then \code{\link{yaml.load}}
on the output file, except that it also supports reading multiple YAML
documents from a single file.
}
