% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{read_test}
\alias{read_test}
\title{Read a test from a file}
\usage{
read_test(
  path,
  allow_multiple = FALSE,
  normalize = TRUE,
  normalise = normalize,
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{allow_multiple}{Whether to allow multiple tests in the file.
These must be separate by the YAML document boundary, \verb{---}. If this
is true, the return value is a list of tests; otherwise, the return
value is a single test.}

\item{normalize}{Whether to normalize the test. This will convert the
test to the canonical format used by the package, which also performs
validation of the test.}

\item{normalise}{Synonym for \code{normalize}.}

\item{envir}{The environment to use for RYaml files.}

\item{...}{Additional arguments to pass to \code{\link[=render_ryaml]{render_ryaml()}}, if it's needed.}

\item{file}{The file to read the test from. Either a YAML file, or an
RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used to render it).}
}
\value{
A test or a list of tests
}
\description{
This function reads a test from a file. The file can be a YAML
file, or an RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used to render it).
}
