% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryaml.R
\name{render_ryaml}
\alias{render_ryaml}
\title{Render an Ryaml file to a YAML file}
\usage{
render_ryaml(
  from,
  to,
  embed_images = TRUE,
  clean_intermediates = embed_images,
  envir = parent.frame(),
  validate = TRUE,
  ...
)
}
\arguments{
\item{from}{The input Ryaml file}

\item{to}{The output YAML file}

\item{embed_images}{Whether to embed images in the output YAML file in
base64 format}

\item{clean_intermediates}{Whether to clean the intermediate files
created by knitr}

\item{envir}{The environment to use for evaluating R code}

\item{validate}{Whether to validate the output YAML file. This is done by
calling \code{\link[yaml]{yaml.load}} on the output file; if an error is
thrown, the function will stop with an error message.}

\item{...}{Additional arguments to pass to \code{\link[knitr]{knit}}}
}
\value{
The output of \code{\link[knitr]{knit}}
}
\description{
This function renders an Ryaml file to a YAML file using the knitr package.
}
\details{
The wrapper around knitr is very thin. The only addition is that you can set
the \code{render} option to \code{yaml_print} to print the output in YAML format;
when paired with echo = FALSE and results = 'asis', this will print the YAML
to the output document. Since knitr maintains the indentation of the output
equal to that of the code chunk, the YAML will be rendered in a way that is
a valid sub-property of where the code chunk is in the document.
}
