% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-to-bbxml.R
\name{to_bbxml.r2bb_question_matching}
\alias{to_bbxml.r2bb_question_matching}
\alias{to_bbxml.r2bb_question_single_blank}
\alias{to_bbxml.r2bb_question_multiple_blanks}
\alias{to_bbxml.r2bb_question_multiple_choice}
\alias{to_bbxml.r2bb_question_multiple_answer}
\alias{to_bbxml.r2bb_question_numeric}
\alias{to_bbxml.r2bb_question_file_upload}
\alias{to_bbxml.r2bb_question_short_answer}
\title{Convert a question to Blackboard XML}
\usage{
\method{to_bbxml}{r2bb_question_matching}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_single_blank}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_multiple_blanks}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_multiple_choice}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_multiple_answer}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_numeric}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_file_upload}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

\method{to_bbxml}{r2bb_question_short_answer}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)
}
\arguments{
\item{x}{The question to convert}

\item{convert_rich_text}{Method to convert rich text to HTML. By default,
\code{md_to_html_pandoc} is used, which converts Markdown to HTML using
\href{https://pandoc.org/}{pandoc}. Can be FALSE to prevent any conversion.}

\item{convert_rich_text_options}{Additional arguments to pass to the
\code{convert_rich_text} function.}

\item{...}{Additional arguments passed to methods}
}
\value{
A character string containing the Blackboard XML
}
\description{
This function converts a question to Blackboard XML format.
}
\seealso{
\code{\link[=to_bbxml]{to_bbxml()}}
}
