% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{to_bbxml.r2bb_test}
\alias{to_bbxml.r2bb_test}
\title{Convert a test to Blackboard XML}
\usage{
\method{to_bbxml}{r2bb_test}(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)
}
\arguments{
\item{x}{The test to convert}

\item{convert_rich_text}{Method to convert rich text to HTML. By default,
\code{md_to_html_pandoc} is used, which converts Markdown to HTML using
\href{https://pandoc.org/}{pandoc}. Can be FALSE to prevent any conversion.}

\item{convert_rich_text_options}{Additional arguments to pass to the
\code{convert_rich_text} function.}

\item{...}{Additional arguments passed to methods}
}
\value{
A character string containing the Blackboard XML
}
\description{
This function converts a test to Blackboard XML format.
}
\seealso{
\code{\link[=to_bbxml]{to_bbxml()}}
}
