% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-bbxml.R
\name{to_bbxml}
\alias{to_bbxml}
\title{Convert an object to Blackboard XML}
\usage{
to_bbxml(x, ...)
}
\arguments{
\item{x}{The object to convert to Blackboard XML}

\item{...}{Additional arguments passed to methods}
}
\value{
A character string containing the Blackboard XML
}
\description{
This function converts an object to Blackboard XML format.
It is a generic function that dispatches based on the class of the object.
}
\seealso{
\code{\link[=to_bbxml.r2bb_question_matching]{to_bbxml.r2bb_question_matching()}}
}
