% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{to_bbxml_package}
\alias{to_bbxml_package}
\title{Convert an R2BB object to a BbXML package}
\usage{
to_bbxml_package(
  x,
  output_path,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(deferred = TRUE),
  postprocess_xml = "md_to_html_pandoc_convert_deferred",
  postprocess_xml_options = list(options = "--mathml")
)
}
\arguments{
\item{x}{An R2BB object to convert}

\item{output_path}{The path to save the output BbXML package}

\item{convert_rich_text}{Whether to convert rich text to HTML. By default,
\code{md_to_html_pandoc_mark_for_conversion} is used, which converts Markdown to
HTML using \href{https://pandoc.org/}{pandoc}. Can be FALSE to prevent any
conversion.}
}
\value{
The path to the output BbXML package
}
\description{
This function converts an R2BB object to a BbXML package, which can be
used to import into Blackboard.
}
