% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{to_markdown.r2bb_pool}
\alias{to_markdown.r2bb_pool}
\title{Convert a question pool to Markdown}
\usage{
\method{to_markdown}{r2bb_pool}(x, ...)
}
\arguments{
\item{x}{The question pool to convert}

\item{...}{Additional arguments passed to methods}
}
\value{
A character string containing the Markdown
}
\description{
This function converts a question pool to Markdown format.
}
\seealso{
\code{\link[=to_markdown.r2bb_question]{to_markdown.r2bb_question()}}
}
