% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-to-markdown.R
\name{to_markdown.r2bb_question}
\alias{to_markdown.r2bb_question}
\alias{to_markdown.r2bb_question_matching}
\alias{to_markdown.r2bb_question_single_blank}
\alias{to_markdown.r2bb_question_multiple_blanks}
\alias{to_markdown.r2bb_question_numeric}
\alias{to_markdown.r2bb_question_file_upload}
\alias{to_markdown.r2bb_question_short_answer}
\alias{to_markdown.r2bb_question_multiple_choice}
\alias{to_markdown.r2bb_question_multiple_answer}
\title{Convert a question to Markdown}
\usage{
\method{to_markdown}{r2bb_question}(x, ...)

\method{to_markdown}{r2bb_question_matching}(x, ...)

\method{to_markdown}{r2bb_question_single_blank}(x, ...)

\method{to_markdown}{r2bb_question_multiple_blanks}(x, ...)

\method{to_markdown}{r2bb_question_numeric}(x, ...)

\method{to_markdown}{r2bb_question_file_upload}(x, ...)

\method{to_markdown}{r2bb_question_short_answer}(x, ...)

\method{to_markdown}{r2bb_question_multiple_choice}(x, ...)

\method{to_markdown}{r2bb_question_multiple_answer}(x, ...)
}
\arguments{
\item{x}{The question to convert}

\item{...}{Additional arguments to pass to the conversion method}
}
\description{
This function is used to convert a question to Markdown format.
}
