% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-markdown.R
\name{to_markdown}
\alias{to_markdown}
\title{Convert an object to Markdown}
\usage{
to_markdown(x, ...)
}
\arguments{
\item{x}{The object to convert to Markdown}

\item{...}{Additional arguments passed to methods}
}
\value{
A character string containing the Markdown
}
\description{
This function converts an object to Markdown format.
It is a generic function that dispatches based on the class of the object.
}
\seealso{
\code{\link[=to_markdown.r2bb_exercise_matching]{to_markdown.r2bb_exercise_matching()}}
}
