% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-read-write.R
\name{write_question}
\alias{write_question}
\title{Write a question to a file}
\usage{
write_question(x, path, ...)
}
\arguments{
\item{x}{The question or questions to write to the file.}

\item{path}{The path to the file to write the question to.}

\item{...}{Additional arguments to pass to \code{\link[=render_ryaml]{render_ryaml()}}, if it's needed.}
}
\description{
This function writes a question to a file. The file can be a YAML file, or an
RYaml file (in which case \code{\link[=render_ryaml]{render_ryaml()}} is used to render it).
}
