% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryaml.R
\name{yaml_print}
\alias{yaml_print}
\title{Print an object in YAML format}
\usage{
yaml_print(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments to pass to \code{yaml::as.yaml}}
}
\description{
This function prints an object in YAML format. It is used by the \code{render_ryaml}
function to print the output in YAML format.
}
